move constants to data structures
