+++ src/main/java/org/gridsuite/modification/dto/byfilter/formula/FormulaInfos.java
@@ -14,6 +14,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 import org.gridsuite.modification.dto.byfilter.AbstractAssignmentInfos;
+import org.gridsuite.modification.modifications.data.assignment.Operator;
+import org.gridsuite.modification.modifications.data.assignment.ReferenceFieldOrValue;
 
 /**
  * @author Seddik Yengui <Seddik.yengui at rte-france.com>
