+++ src/main/java/org/gridsuite/modification/modifications/byfilter/ByFormulaModification.java
@@ -17,8 +17,8 @@ import lombok.*;
 import org.gridsuite.modification.ModificationType;
 import org.gridsuite.modification.dto.byfilter.AbstractAssignmentInfos;
 import org.gridsuite.modification.dto.byfilter.formula.FormulaInfos;
-import org.gridsuite.modification.dto.byfilter.formula.Operator;
 import org.gridsuite.modification.error.NetworkModificationExceptionType;
+import org.gridsuite.modification.modifications.data.assignment.Operator;
 import org.gridsuite.modification.report.NetworkModificationReportResourceBundle;
 
 import java.math.BigDecimal;
