+++ src/main/java/org/gridsuite/modification/modifications/data/assignment/FormulaAssignmentData.java
@@ -13,8 +13,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.gridsuite.modification.dto.byfilter.formula.Operator;
-import org.gridsuite.modification.dto.byfilter.formula.ReferenceFieldOrValue;
 
 /**
  * @author Seddik Yengui <Seddik.yengui at rte-france.com>
