+++ src/main/java/org/gridsuite/modification/modifications/data/assignment/ReferenceFieldOrValue.java
@@ -1,11 +1,11 @@
-/**
- * Copyright (c) 2023, RTE (http://www.rte-france.com)
+/*
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.gridsuite.modification.dto.byfilter.formula;
+package org.gridsuite.modification.modifications.data.assignment;
 
 import com.powsybl.iidm.network.Identifiable;
 import lombok.*;
