+++ src/main/java/org/gridsuite/modification/modifications/data/assignment/ValueAssignmentData.java
@@ -17,7 +17,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.gridsuite.modification.dto.byfilter.assignment.*;
 
 /**
  * @author Joris MANCINI <joris.mancini at rte-france.com>
