+++ src/test/java/org/gridsuite/modification/modifications/byfilter/formula/GeneratorByFormulaModificationTest.java
@@ -15,8 +15,8 @@ import org.gridsuite.modification.dto.FilterEquipments;
 import org.gridsuite.modification.dto.IdentifiableAttributes;
 import org.gridsuite.modification.dto.byfilter.equipmentfield.GeneratorField;
 import org.gridsuite.modification.dto.byfilter.formula.FormulaInfos;
-import org.gridsuite.modification.dto.byfilter.formula.Operator;
-import org.gridsuite.modification.dto.byfilter.formula.ReferenceFieldOrValue;
+import org.gridsuite.modification.modifications.data.assignment.Operator;
+import org.gridsuite.modification.modifications.data.assignment.ReferenceFieldOrValue;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
